CodingStyle: Run indent over the .cpp files
The first thing we need to do before starting any refactoring work is to
get a consistent coding style accross the source.

For this I've used the Linux kernels Lindent script which is a wrapper
around the indent[0] program. This has put the source into a coding
style[1] as used by the Linux kernel, which I personally find produces
the most readable code.

This should provide a good base for further tweaking.

[0] - http://www.gnu.org/software/indent/
[1] -
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/tree/Documentation/CodingStyle

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>