Dynamically allocate various data structures
signalserver has two modes of operation; 90m and HD (30m). Depending on
which mode is being used various arrays are sized differently.

Previously this was done by having separate binaries for the two modes
with a #ifdef in the code that would size stuff differently depending on
whather -DHD was passed to the compiler.

What this commit does is get rid of the #ifdef and dynamically size the
arrays at run-time and only produce a single binary. We select which to
use by either calling signalserver or signalserverHD (a symlink to
signalserver) and checking argv[0].

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>