Silence a compiler warning
We were getting the following

    inputs.cc: In function ‘int LoadSDF_SDF(char*, int)’:
    inputs.cc:120:33: warning: ‘data’ may be used uninitialized in this
    function [-Wmaybe-uninitialized]
        dem[indx].data[x][y] = data;

data is set by converting values read in from a file to an integer,
this bit is guarded by an if (fgets() != NULL) this is done in a loop.
It's possible that the first time round the loop (or every time round
the loop) data could be used uninitialised. The other case could be that
previous values of data could be used if the value failed to be read
from the file.

I'm guessing that these cases shouldn't generally happen so the simplest
solution is to just initialise data to 0.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>