main.cc: Remove the build and gcc version info
Remove the comment near the top of the file that spoke about the version
of GCC required and g++ comamnd to use to build it.

A GCC version check has been added to the Makefile and rather than
update the g++ command it's probably best to just remove it, we have
make!

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>