models: hata: Remove HATApathLoss() from models.cc
This function was already in a hata.cc file, although called HataLinkdB,
rename it to HATApathLoss as that's what it's called as from elsewhere.

Also remove the functions; incidenceAngle & ked from hata.cc file as
they are not used here and are still in the models.cc file (will be
moved in another commit).

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>